separate dt2sql
